Adjust to restricted_access doesnt delete others admins

+++ app/controllers/backoffice/admins_controller.rb
@@ -1,7 +1,7 @@
 class Backoffice::AdminsController < BackofficeController
   before_action :set_admin, only: [:edit, :update, :destroy]
 
-  after_action :verify_authorized, only: :new
+  after_action :verify_authorized, only: [:new, :destroy]
   after_action :verify_policy_scoped, only: :index
 
   def index
@@ -35,10 +35,11 @@ class Backoffice::AdminsController < BackofficeController
   end
 
   def destroy
-    admin_email = @admin.email
+    authorize @admin
+    admin_name = @admin.name
 
     if @admin.destroy
-      redirect_to backoffice_admins_path, notice: "O administrador #{admin_email} foi apagado com sucesso!"
+      redirect_to backoffice_admins_path, notice: "O administrador #{admin_name} foi apagado com sucesso!"
     else
       render :index
     end
